+++ src/execution/CommonExecutionToolsOptions.ts
@@ -1,9 +1,9 @@
-export interface CommonExecutionToolsOptions {
+export type CommonExecutionToolsOptions = {
     /**
      * If true, the internal executions will be logged
      */
     readonly isVerbose?: boolean;
-}
+};
 
 /**
  * TODO: [🧠][💙] Distinct between options passed into ExecutionTools and to ExecutionTools.execute