+++ src/execution/translation/automatic-translate/automatic-translators/ITranslatorOptions.ts
@@ -1,4 +1,0 @@
-export interface TranslatorOptions {
-    from: string;
-    to: string;
-}