+++ src/execution/translation/automatic-translate/automatic-translators/TranslatorOptions.ts
@@ -1,0 +1,4 @@
+export type TranslatorOptions = {
+    from: string;
+    to: string;
+};