+++ src/types/Command.ts
@@ -23,10 +23,10 @@ export type Command =
  * - It is used for backward compatibility
  * - It is defined per whole .promptbook file in the header
  */
-export interface PromptbookUrlCommand {
+export type PromptbookUrlCommand = {
     readonly type: 'PROMPTBOOK_URL';
     readonly promptbookUrl: URL;
-}
+};
 
 /**
  * PromptbookVersion command tells which version is .promptbook file using
@@ -34,37 +34,37 @@ export interface PromptbookUrlCommand {
  * - It is used for backward compatibility
  * - It is defined per whole .promptbook file in the header
  */
-export interface PromptbookVersionCommand {
+export type PromptbookVersionCommand = {
     readonly type: 'PROMPTBOOK_VERSION';
     readonly promptbookVersion: string_version;
-}
+};
 
 /**
  * Execute command tells how to execute the section
  * It can be either prompt template, script or SIMPLE TEMPLATE etc.
  */
-export interface ExecuteCommand {
+export type ExecuteCommand = {
     readonly type: 'EXECUTE';
     readonly executionType: ExecutionType;
-}
+};
 
 /**
  * Model command tells which model and modelRequirements to use for the prompt template. execution
  */
-export interface ModelCommand {
+export type ModelCommand = {
     readonly type: 'MODEL';
     readonly key: keyof ModelRequirements;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     readonly value: any /* <- TODO: Infer from used key, can it be done in TypeScript */;
-}
+};
 
 /**
  * Joker parameter is used instead of executing the prompt template if it meet the expectations requirements
  */
-export interface JokerCommand {
+export type JokerCommand = {
     readonly type: 'JOKER';
     readonly parameterName: string_name;
-}
+};
 
 /**
  * Parameter command describes one parameter of the prompt template
@@ -73,22 +73,22 @@ export interface JokerCommand {
  * - It can have description
  * - In description it can have simple formatting BUT not markdown structure or reference to other parameters
  */
-export interface ParameterCommand {
+export type ParameterCommand = {
     readonly type: 'PARAMETER';
     readonly isInput: boolean;
     readonly isOutput: boolean;
     readonly parameterName: string_name;
     readonly parameterDescription: string_markdown_text | null;
-}
+};
 
 /**
  * Postprocess command describes which function to use for postprocessing
  * This will be created as separate EXECUTE SCRIPT block bellow
  */
-export interface PostprocessCommand {
+export type PostprocessCommand = {
     readonly type: 'POSTPROCESS';
     readonly functionName: string_name;
-}
+};
 
 /**
  * Expect command describes the desired output of the prompt template (after post-processing)
@@ -102,17 +102,17 @@ export type ExpectCommand = ExpectAmountCommand | ExpectFormatCommand;
  *
  * Note: LLMs work with tokens, not characters, but in Promptbooks we want to use some human-recognisable and cross-model interoperable units.
  */
-export interface ExpectAmountCommand {
+export type ExpectAmountCommand = {
     readonly type: 'EXPECT_AMOUNT';
     readonly sign: 'EXACTLY' | 'MINIMUM' | 'MAXIMUM';
     readonly unit: ExpectationUnit;
     readonly amount: ExpectationAmount;
-}
+};
 
 /**
  * Represents a command that expects a specific format.
  */
-export interface ExpectFormatCommand {
+export type ExpectFormatCommand = {
     readonly type: 'EXPECT_FORMAT';
     readonly format: 'JSON';
-}
+};